chore: extract `SecurityDetails` into its own module
